Move main App component up one level. Reorganize a few imports
