+++ src/App/NoteDemo/PlayRecordButton/PlayRecordButton.js
