+++ src/App/NoteDemo/PlayRecordButton/index.js
