+++ src/App/NoteDemo/ReplayVisibilityWrapper/DownloadLink/DownloadLink.js
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 export default class DownloadLink extends Component {
 
   render() {
-
     return (
       <a href={this.props.url}
         download={this.props.demoName.toLowerCase() + '_web_audio_demo.wav'}
