+++ src/App/NoteDemo/ReplayVisibilityWrapper/ReplayWidget/ReplayWidget.js
