--- src/components/index.js
@@ -1,7 +1,0 @@
-import App from './App/App';
-import NoteDemo from './NoteDemo/NoteDemo';
-
-export {
-  App,
-  NoteDemo
-};
