+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+
+import App from './App';
 import './index.css';
-import App from 'components/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
